refactor the tabs menu components in small pieces
