--- client/src/components/NavBar/components/TabsMenu.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const TabsMenu = () => {
-  return <div>TabsMenu</div>;
-};
-
-export default TabsMenu;
