+++ client/src/components/NavBar/components/tabsMenu/HomeTab.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const HomeTab = () => {
+  return (
+    <button
+      type="button"
+      href="#"
+      className="w-full text-3xl py-2 px-3 xl:px-12 cursor-pointer text-center inline-block text-blue-500 border-b-4 border-blue-500"
+    >
+      <i className="bx bxs-home" />
+    </button>
+  );
+};
+
+export default HomeTab;
