+++ client/src/components/NavBar/components/tabsMenu/MarketplaceTab.jsx
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const Marketplace = () => {
+  return (
+    <button
+      type="button"
+      href="#"
+      className="w-full text-3xl py-2 px-3 xl:px-12 cursor-pointer text-center inline-block rounded text-gray-600 hover:bg-gray-100 dark:hover:bg-dark-third dark:text-dark-txt relative"
+    >
+      <i className="bx bx-store" />
+    </button>
+  );
+};
+
+export default Marketplace;
